UISINVCOMP-75 Allow tenant to set default columns to display in Inventory results.
