+++ lib/constants/searchResultsList.js
@@ -1,13 +1,6 @@
 import { FormattedMessage } from 'react-intl';
 
-import { SORT_OPTIONS } from './constants';
-
-export const SEARCH_COLUMN_NAMES = {
-  TITLE: SORT_OPTIONS.TITLE,
-  CONTRIBUTORS: SORT_OPTIONS.CONTRIBUTORS,
-  PUBLISHERS: 'publishers',
-  DATE: SORT_OPTIONS.DATE,
-};
+import { SEARCH_COLUMN_NAMES } from './constants';
 
 export const SEARCH_COLUMN_MAPPINGS = {
   [SEARCH_COLUMN_NAMES.TITLE]: <FormattedMessage id="stripes-inventory-components.instances.columns.title" />,
